requirements.txt
flask
deep_translator
scikit-learn==1.3.1
pymongo
flask_cors
dnspython
python-dotenv
tweepy